Adopt @openshift/dynamic-plugin-sdk as the dynamic plugin SDK
Since the @openshift-console/dynamic-plugin-sdk (console SDK) is
compatible with and will be integrating with the
@openshift/dynamic-plugin-sdk (core SDK), adopt the use of the core
SDK for basic webpack builds and the project's console dynamic plugin
configurations.

TypeScript configs are updated to follow the settings in the core SDK.
The only big difference is that strict mode is left off.

Yarn scripts are updated taking into consideration the TypeScript
configs.

Created `DynamicConsoleRemotePlugin` webpack plugin that modifies the
work of the core SDK webpack plugins to be compatible with console SDK
4.11 runtime.  Shared module handling to match console is currently
embedded in the repo as `src/webpack/shared-modules.ts`.  In future
core SDK releases, some of the needed variables may be accessible
directly from the SDK.

Configuration changes:

  - Move package.json `.consolePlugin` to `plugin.json`

  - Console extensions now defined in `console-extensions.ts` allowing
    static type checking of the extensions

  - webpack dev and prod configs are embedded in same structure

  - webpack output filenames updated to be directory based

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>