Convert `plugin.json` to `plugin.ts`
By changing the plugin definition from `json` to `ts` we gain the
ability to break up the `exposedModules` definitions however is most
convient for us.

The custom webpack plugin `DynamicConsoleRemotePlugin` was modified
to support reading plugin metadata from a JSON file or to accept it
as an already parsed object.  This follows the same pattern used for
the extensions array, and core sdk's plugin metadata handling
capabilities.

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>