Fix eslint for ci and ide on package/legacy
eslint with typescript in a monorepo has a few quirks.  One big one is
having to set `parserOptions.tsconfigRootDir` such that eslint works
from CI/CLI and within an IDE (vscode with eslint extension).  The
workaround is to use a javascript config file and the directive
`__dirname` to pull the file's actual directory name.  With that in
place, vscode linting works again!

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>