Drop use of the `Q.allSettled()`
The old Q promise library was being used solely for the `allSettled()`
function [1].  This allows multiple async function / promises to run in
parallel and collect the full results of each.  Use of the Q library
just for this function is no longer necessary.  The standard `Promise`
object contains the same `allSettled()` function with the same
signiture [2].

[1] - https://github.com/kriskowal/q#combination
[2] - https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Promise/allSettled

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>