Require namespace for routed sub-pages
Fix problem with name collisions between resources from different
namespaces.

Before, the namespace provided via the URL referred to workspace
namespace (chosen by the user). This allowed to pass that value to
legacy code which is not aware of console global state (which namespace
is part of).
After the fix, the namespace always refers to the resource namespace.
If the namespace is unknown (create operation in All Namespaces view)
then the ENV.DEFAULT_NAMESPACE is used.

Side effect of this fix is that the previous workspace namespace is
no longer preserved. Example:
1. choose 'All Namespaces' on Plans list screen
2. enter create plan screen
3. cancel create action
4. the namespace on Plans list screen is now 'konveyor-forklift'
   (or other defined at compile time using ENV.DEFAULT_NAMESPACE)

Signed-off-by: Radoslaw Szwajkowski <rszwajko@redhat.com>