If insecure mode is set for a provider, avoid cacert field validation
Fix a bug for cases in which the insecure mode is set for a provider's
create/edit dialog and the 'CA certificate' field is set with an invalid
value. This is relevant only for oVirt, OpenStack providers.

Before this fix: the validation failed the whole form and therefore the
Create/Save buttons were disabled (even though the CA certificate is
hidden and not relevant for the insecure mode).

After this fix: the validation succeded and the Create/Save buttons are
enabled. Once the secure mode is choosen again, the original invalid 'CA certificate'
field value is displayed and Create/Save buttons are disabled.

Signed-off-by: Sharon Gratch <sgratch@redhat.com>