Setup mocks via MockServiceWorker as a optional console extension
Setup mock data as external to the query code using MockServerWorker
(MSW).  This configuration allows the query code to have minimal
knowledge of if the plugin is running on mock data.  Production builds
will not need to include mock data.

Use a `console.context-provider` console extension point to inject
the mock service worker very close to the start up of console.  The
MSW worker is only setup once.  All setup handlers are logged to the
console before the context provider is done running.  The MSW worker
reference is saved to `window.msw` to help with dev/debug.

The console extension and MSW is provided in `forklift-console-plugin`
and uses MSW rest providers from the `mocks` package.  The extension
is only included in the build if `DATA_SOURCE=mock`.

To run the plugin in dev mode with mock data, the same command still
works:
```sh
> DATA_SOURCE=mock npm run start
```

Extension specifics:
  - webpack dev server setup to serve MSW's `mockServerWorker.ts`
  - webpack dev server adds response header `Service-Worker-Allowed`
    with the value `/`.  Setting the header to `/` allows MSW to
    scope to `/` and mock ALL calls.
  - `forklift-console-plugin` has new extension `mock-console-extension`
    and it's `dynamic-plugin.ts` is sensitive to the `DATA_SOURCE`
    environment variable.

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>