Fix import paths used for monorepo packages
Some files in `forklift-console-plugin` were using the wrong import
paths to reference other packages in the monorepo.  The reason they
still worked was the extra paths references needed in the plugin's
`tsconfig.json` to handle the path references in the legacy and common
packages.

The ultimate goal is to be able to remove the need for config kludges
like this (tsconfig.json path references across projects) and have
each package be directly consumable.

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>