Create `BuildMetadataPlugin` to add a `build-metadata.json` file to the build
In a new customer webpack plugin, collect various build time information,
primarily from `package.json` and the git repo, and push it to the build
`dist/` folder.  The info is included in 2 forms:
  - banner comments in each output webpack chunk
  - `build-metadata.json` in the style of `@openshift/dynamic-plugin-sdk`

Both of these provide helpful information to anyone looking to trace
back the source of the built code.

The `build-metadata.json` will be easy to access from the runtime
container with a podman command like:
```sh
  podman run -i -t --rm --pull=always \
    quay.io/kubev2v/forklift-console-plugin:latest \
    cat build-metadata.json
```

For `getBuildMetadata()` and `getBaseBuildMetadata()`, replaced the
`buildDate` and `buildTime` fields with the single `buildDateTime`
value.  The new value is Date.toISOString() / ISO 8601 formatted to
be stable across any system timezone or locale settings.

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>