ci: cope with tls-enforcal for forklift-inventory
These changes allow console in dev-mode to work with forklift-inventory
service that is set with TLS enabled: we connect to the service using
`kubectl port-forward` and change the default port from 30088 to 65300.

There is a room for cleanup in the scripts that spin up a kind cluster
but that's out of the scope of these changes.

Signed-off-by: Arik Hadas <ahadas@redhat.com>