Generically config eslint to include all relevant file extensions
With multiple file extensions around the monorepo, it is easier to
setup the eslint config plugin to force include all of the file
extensions we care about.  This is done, in a very hacky way, using an
`overrides` section to layer inclusions on top of the base directory.
The `--ext` extension is no longer needed.

File extensions we care about linting:
  - .js
  - .jsx
  - .cjs
  - .mjs
  - .ts
  - .tsx

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>