Fix eslint on the legacy package
With commit 62023c6 moving the eslint config file for the legacy
package from the generic `.eslintrc` json file to `.eslintrc.js`,
the configuration was broken.  The new js file did not properly
export the config using `module.export` instead of the correct
`module.exports`.

Fixing the export to be `module.exports` and adopting the `ignorePatterns`
and `overrides` sections from #486 and #495 fixes the linting and
brings the included/excluded files inline with the rest of the
monorepo's linting.

Linting errors are present (315 errors,4 warnings).  They will be fixed
in a following commit.

Signed-off-by: Scott J Dickerson <sdickers@redhat.com>