Rerender the Mapping actions dropdown when permissions are changed
This fixes a bug of Mappings actions dropdown menu options that are always disabled.

As a followup for
https://github.com/kubev2v/forklift-console-plugin/pull/636, we now need
to re-render the Storage/Network Mappings actions dropdown menu once the
user's permissions are changed.

The root cause is that since the user's access permissions are set to false at loading till it is synced,
the menu options are disabled at start and need to be re-rendered once
the data is synced.

Signed-off-by: Sharon Gratch <sgratch@redhat.com>