Enhance EditModal to refresh the helperText field value based on validationHook result
For supporting refreshing the helperText field value displayed, based on the
ValidationResults (e.g., success, warning), need to set it to the
validation.helperText value instead of the constant helperText prop
value.

If validation.helperText is set to '' (initial state or for backward compatible use),
then use the initial helperText prop as was done before this fix.

Signed-off-by: Sharon Gratch <sgratch@redhat.com>