Display the "x" icon in the Overview ShowWelcomeCardButton button
To be consistent with the appearance on the console Home->Overview page,
display the "x" (close) icon in the ShowWelcomeCardButton button label.
The ShowWelcomeCardButton button label appears in the Overview page when
the Welcome card is hidden

Note: in comparing to the console Home->Overview page, the "x" action
is disabled, so the button won't be closed/vanised as a result of clicking it.

Signed-off-by: Sharon Gratch <sgratch@redhat.com>