Enhance the DetailsItem component to enable both hyperlink and edit
Reference: https://github.com/kubev2v/forklift-console-plugin/issues/1106

Up till now the DetailsItem component could support 2 modes:
1. A non editable content (w/o hyperlink).
2. An editable content while both content and pencil icon are inline button for
triggering the onEdit callback.

This PR enhances the DetailsItem component to replace the editable mode (2) style such that
the field context is an element (can be a hyperlink button for example) and only the pencil icon
is an inline button for triggering the onEdit callback.

We can consider adding a 'Edit' label next to the pencil icon.

Signed-off-by: Sharon Gratch <sgratch@redhat.com>