Fix Negative number for plan/migration status's VMs count
On both plan and migration status fields, the number of displayed
migrated VMs within the status' progress bar label is sometimes negative,
i.e.  the value of X within "X out of Y VMs migrated" is sometimes
negative.

This is reproduced when at least one VM failed to migrated.

The reason is that failed VMs is often marked with an error but their completed
value is still set to false since VM's migration was not finished yet. So the
calculation of migrated = completed - error < 0.

Signed-off-by: Sharon Gratch <sgratch@redhat.com>