Add UI for setting inventory memory limits
There are currently knobs that can be used to adjust the memory limit
for the `main` container, but not for the `inventory` container. This
can cause the inventory container to go into a crash loop on a large
deployment. Add the ability to bump the inventory memory limit  to avoid
this.

Fixes: https://issues.redhat.com/browse/MTV-1504

Signed-off-by: Jonathon Jongsma <jjongsma@redhat.com>