When migration task isn't completed but the migration task progress.total equals progress.completed, mark task as successful
Reference: https://issues.redhat.com/browse/MTV-1718

In case a VM migration task is not marked as finished (i.e. task.phase is undefined),
and its pipeline step is not marked as completed as well (but not failed) then mark the task in UI as completed successfully if taks's progress total reaches the task's progress completed.

This will solve cases in which the DiskTransferV2v or DiskTransfer tasks aren't completed although the migration is successful.

This fix was done for all migration VM types.

Signed-off-by: Sharon Gratch <sgratch@redhat.com>