Fix StorageMaps, NetworkMaps pages for referring to mapping's source/destination providers based on provider names
Reference: https://issues.redhat.com/browse/MTV-1421

Currently when displaying a storage mappings or metwork mappings in StorageMaps or NetworkMaps pages,
the reference to the mapping's source/destination providers was done
based on providers uid only and not also on providers name+namepsace.

That was problematic when loading a mappings via CL yaml since the
yaml required format doesn't include providers uids data. So when loading
mappings via a yaml, the uids for the providers should have been set or
otherwise the mappings were wrong displayed and weren't editable.

Note that mapppings displayed in the plan are supported for both name or
uid.

Examples:
This yaml format didn't work:

kind: NetworkMap
...
spec:
   provider:
     source:
       name: vsphere
       namespace: konveyor-forklift
     destination:
      name: host
      namespace: konveyor-forklift

While this worked:

kind: NetworkMap
...
spec:
   provider:
     source:
       name: vsphere
       namespace: konveyor-forklift
       uid: 6cf88e38-063c-4cee-9240-cc02e023157a
     destination:
      name: host
      namespace: konveyor-forklift
      uid: 9562477b-0297-45fc-b5d4-7f61de98a002

This PR fixes that so both yaml formats are acceptable.

Signed-off-by: Sharon Gratch <sgratch@redhat.com>