Disabling all hooks tab fields in case hooks editing is not allowed
Reference: https://issues.redhat.com/browse/MTV-1738

A follow up fix for disabling all hooks tab fields in case the plan
status is either 'archived' or 'running'.

The only exception is that the CodeEditor component can't be disabled
for entering the Ansible playbook content, but I think it's ok to leave
it as is for now since no upadting action can be done.

Signed-off-by: Sharon Gratch <sgratch@redhat.com>