MTV-2971: display the 'clear all filters' button for WwitchFilter component
Reference: https://issues.redhat.com/browse/MTV-2971

When a filter from type of SwitchFilter is applied. the 'clear all filters'
button is not displayed, even though clicking on this button (once displayed for other filters)
does reset the switch filter to default value.

An example for that is the plans list -> 'show archived' persistent filter.

Solving it by adding a chip for the SwitchFilter which will enable
adding the reset button to the toolbar. Once clicking on the reset
button, all filters including the 'show archived' are now reset to
default.

Signed-off-by: Sharon Gratch <sgratch@redhat.com>