MTV-3372: add support for uploading a local OVA file for an OVA provider (#1828)
Reference: https://issues.redhat.com/browse/MTV-3372

Add an option to upload local OVA files (one by one) for an OVA
provider.

This PR includes only the UI components since the backend api for that
is still not ready at that point. Curenttly there is a call to a dummy
api that should be replaced with the correct one.

Signed-off-by: Sharon Gratch <sgratch@redhat.com>